+++ jupyterthemes/stylefx.py
@@ -1,12 +1,11 @@
 from __future__ import print_function
-
 import os
 import sys
 from shutil import copyfile, rmtree
-
 from jupyter_core.paths import jupyter_config_dir, jupyter_data_dir
-
 import lesscpy
+from IPython.core.display import HTML
+from glob import glob
 
 # path to local site-packages/jupyterthemes
 package_dir = os.path.dirname(os.path.realpath(__file__))
@@ -41,6 +40,7 @@ ex_style = os.path.join(layouts_dir, 'extras.less')
 jax_style = os.path.join(layouts_dir, 'mathjax.css')
 vim_style = os.path.join(layouts_dir, 'vim.less')
 comp_style = os.path.join(layouts_dir, 'completer.less')
+theme_name_file = os.path.join(jupyter_custom, 'current_theme.txt')
 
 
 def check_directories():
@@ -55,7 +55,6 @@ def check_directories():
         os.makedirs(jupyter_data)
     if not os.path.isdir(jupyter_nbext):
         os.makedirs(jupyter_nbext)
-
 
 def less_to_css(style_less):
     """ write less-compiled css file to jupyter_customcss in jupyter_dir
@@ -94,7 +93,7 @@ def install_precompiled_theme(theme):
     copyfile(theme_src, theme_dst)
 
     for fontcode in ['exosans', 'loraserif', 'droidmono', 'firacode']:
-        fname, fpath, ffam = stored_font_dicts(fontcode)
+        fname, fpath = stored_font_dicts(fontcode)
 
         fontpath = os.path.join(fonts_dir, fpath)
         for fontfile in os.listdir(fontpath):
@@ -112,22 +111,15 @@ def delete_font_files():
         os.remove(abspath)
 
 
-def import_stored_fonts(fontcodes=('exosans', 'loraserif', 'droidmono')):
+def import_stored_fonts(style_less, fontcodes=('exosans', 'loraserif', 'droidmono')):
     """Collect fontnames and local pointers to fontfiles in custom dir
-    then pass information for each font to function for writing import 
+    then pass information for each font to function for writing import
     statements
     """
-    doc = '\nConcatenated font imports, .less styles, & custom variables\n'
-    s = '*' * 65
-    style_less = '\n'.join(['/*', s, s, doc, s, s, '*/'])
-    style_less += '\n\n\n'
-    style_less += '/* Import Notebook, Markdown, & Code Fonts */\n'
-
     for fontcode in set(fontcodes):
-        fname, fpath, ffam = stored_font_dicts(fontcode)
+        fname, fpath = stored_font_dicts(fontcode)
         style_less = import_fonts(style_less, fname, fpath)
     style_less += '\n\n'
-
     return style_less
 
 
@@ -139,37 +131,42 @@ def convert_fontsizes(fontsizes):
             fontsizes[i] = '.'.join([fs[:-1], fs[-1]])
         elif int(fs) > 50:
             fontsizes[i] = '.'.join([fs[0], fs[-1]])
-
     return fontsizes
 
 
-def set_font_properties(nbfont='exosans',
+def set_font_properties(style_less, nbfont='exosans',
                         tcfont='loraserif',
                         monofont='droidmono',
                         monosize=11,
                         tcfontsize=13,
                         nbfontsize=13,
-                        prfontsize=95):
-    """Parent function for setting notebook, text/md, and 
+                        prfontsize=95,
+                        dfonts=False):
+    """Parent function for setting notebook, text/md, and
     codecell font-properties
     """
+
     fontsizes = [monosize, nbfontsize, tcfontsize, prfontsize]
     monosize, nbfontsize, tcfontsize, prfontsize = convert_fontsizes(fontsizes)
-    style_less = import_stored_fonts(
-        fontcodes=[nbfont, tcfont, monofont, 'firacode'])
+
+    if dfonts:
+        monofont = 'monospace'
+        tcfont = 'sans-serif'
+        nbfont = 'sans-serif'
+    else:
+        style_less = import_stored_fonts(style_less,
+            fontcodes=[nbfont, tcfont, monofont, 'firacode'])
+        # get fontname, fontpath, font-family info
+        nbfont, nbfontpath= stored_font_dicts(nbfont)
+        tcfont, tcfontpath= stored_font_dicts(tcfont)
+        monofont, monofontpath = stored_font_dicts(monofont)
+
     style_less += '/* Set Font-Type and Font-Size Variables  */\n'
 
-    # get fontname, fontpath, font-family info
-    nbfont, nbfontpath, nbfontfam = stored_font_dicts(nbfont)
-    tcfont, tcfontpath, tcfontfam = stored_font_dicts(tcfont)
-    monofont, monofontpath, monofontfam = stored_font_dicts(monofont)
-
     # font names and fontfamily info for codecells, notebook & textcells
-    style_less += '@monofont: "{}"; \n'.format(monofont)
-    style_less += '@notebook-fontfamily: "{}", {}; \n'.format(nbfont,
-                                                              nbfontfam)
-    style_less += '@text-cell-fontfamily: "{}", {}; \n'.format(tcfont,
-                                                               tcfontfam)
+    style_less += '@monofont: {}; \n'.format(monofont)
+    style_less += '@notebook-fontfamily: {}; \n'.format(nbfont)
+    style_less += '@text-cell-fontfamily: {}; \n'.format(tcfont)
 
     # font size for codecells, main notebook, notebook-sub, & textcells
     style_less += '@monofontsize: {}pt; \n'.format(monosize)
@@ -195,10 +192,10 @@ def import_fonts(style_less, fontname, font_subdir):
                   'otf': 'opentype'}
 
     define_font = (
-        "@font-face {{font-family: '{fontname}';\n\tfont-weight:"
+        "@font-face {{font-family: {fontname};\n\tfont-weight:"
         "{weight};\n\tfont-style: {style};\n\tsrc: local('{fontname}'),"
         "\n\turl('fonts{sepp}{fontfile}') format('{ftype}');}}\n")
-
+    fontname = fontname.split(',')[0]
     fontpath = os.path.join(fonts_dir, font_subdir)
     for fontfile in os.listdir(fontpath):
         if '.txt' in fontfile or 'DS_' in fontfile:
@@ -237,9 +234,9 @@ def style_layout(style_less,
                  hideprompt=False):
     """Set general layout and style properties of text and code cells"""
 
-    # grade3's altlayout is reverse of default
-    if theme == 'grade3':
-        altlayout = not altlayout
+    # write theme name to ~/.jupyter/custom/ (referenced by jtplot.py)
+    with open(os.path.join(jupyter_custom, 'current_theme.txt'), 'w') as f:
+        f.write(theme)
 
     if (os.path.isdir(styles_dir_user) and
             '{}.less'.format(theme) in os.listdir(styles_dir_user)):
@@ -256,9 +253,13 @@ def style_layout(style_less,
     promptBG = '@cc-input-bg'
     promptPadding = '.25em'
     promptBorder = '2px solid @prompt-line'
-    tcPromptBorder = promptBorder
+    tcPromptBorder = '2px solid @tc-prompt-std'
     promptMinWidth = 12
     tcPromptWidth = promptMinWidth
+
+    # grade3's altlayout is reverse of default
+    if theme == 'grade3':
+        textcell_bg = '@notebook-bg'
     if altprompt:
         promptPadding = '.1em'
         promptMinWidth = 8
@@ -267,8 +268,7 @@ def style_layout(style_less,
     if altlayout:
         # alt txt/md layout
         textcell_bg = '@notebook-bg'
-        tcPromptBorder = '2px solid transparent'
-        tcPromptWidth = 0
+        promptMinWidth = 8
     if margins != 'auto':
         margins = '{}px'.format(margins)
     if '%' not in cellwidth:
@@ -398,10 +398,6 @@ def reset_default(verbose=False):
 def set_nb_theme(name):
     """Set theme from within notebook """
 
-    # TODO move this import to the top?
-    from IPython.core.display import HTML
-    from glob import glob
-
     styles_dir = os.path.join(package_dir, 'styles/compiled/')
     css_path = glob('{0}/{1}.css'.format(styles_dir, name))[0]
     customcss = open(css_path, "r").read()
@@ -411,8 +407,8 @@ def set_nb_theme(name):
 
 def get_colors(theme='grade3', c='default', get_dict=False):
     if theme == 'grade3':
-        cdict = {'b': '#1e70c7',
-                 'default': '#ff711a',
+        cdict = {'default': '#ff711a',
+                 'b': '#1e70c7',
                  'o': '#ff711a',
                  'r': '#e22978',
                  'p': '#AA22FF',
@@ -424,6 +420,9 @@ def get_colors(theme='grade3', c='default', get_dict=False):
                  'r': '#DB797C',
                  'p': '#c776df',
                  'g': '#94c273'}
+
+    cdict['x'] = '@cc-input-fg'
+
     if get_dict:
         return cdict
 
@@ -456,6 +455,7 @@ def stored_font_dicts(fontcode, get_all=False):
               'firacode': ['Fira Code', 'firacode'],
               'generic': ['Generic Mono', 'generic'],
               'hack': ['Hack', 'hack'],
+              'hasklig': ['Hasklig', 'hasklig'],
               'inputmono': ['Input Mono', 'inputmono'],
               'inconsolata': ['Inconsolata-g', 'inconsolata-g'],
               'liberation': ['Liberation Mono', 'liberation'],
@@ -527,8 +527,7 @@ def stored_font_dicts(fontcode, get_all=False):
         fontname, fontdir = fonts['serif'][fontcode]
         fontfam = 'serif'
     else:
-        "One of the fonts you requested is not available... sorry!"
-        return _
+        print("One of the fonts you requested is not available... sorry!")
+        return '', ''
     fontdir = os.sep.join([fontfam, fontdir])
-
-    return fontname, fontdir, fontfam
+    return '"{}", {}'.format(fontname, fontfam), fontdir